Fix #21: Compatibility issue with ffmpeg 0.8
In 0.0.7 --overwrite has been introduced. It adds "-y" or "-n" to the
ffmpeg command to allow to overwrite an already existing file or exit
with error otherwise.

Unfortunately this option has been added quite recently to ffmpeg and
introduce a regression on Ubuntu 12.04, 13.10 and most likely lot of
other stable distros. This patch use "-y" and replace "-n" by a manual
check. A race condition can still occur but it is good enough.