Get rid of the PIL dependency
PIL was used to convert PNG into JPEG files to feed ffmpeg because older releases
of ffmpeg had issue with PNG.

It seems now OK to accept PNG too. Affected user can download a static linked
version of ffmpeg as a workaround.

Removing PIL is important because it prevented the port of infoqscraper to python 3.